data base seeding

+++ nextjs-dashboard/scripts/seed.js
@@ -55,7 +55,7 @@ async function seedInvoices(client) {
     CREATE TABLE IF NOT EXISTS invoices (
     id UUID DEFAULT uuid_generate_v4() PRIMARY KEY,
     customer_id UUID NOT NULL,
-    amount INT NOT NULL,
+    amount BIGINT NOT NULL,
     status VARCHAR(255) NOT NULL,
     date DATE NOT NULL
   );
